Refactor world module into package
