+++ tests/test_coords.py
@@ -1,5 +1,5 @@
 from constants import REGION_SIZE
-from runepy.world import world_to_region, local_tile
+from runepy.world.region import world_to_region, local_tile
 
 
 def round_trip(x, y):
