+++ tests/test_region_bin.py
@@ -2,7 +2,7 @@ import gzip
 from pathlib import Path
 
 import numpy as np
-from runepy.world import Region
+from runepy.world.region import Region
 from constants import REGION_SIZE
