+++ tests/test_region_manager.py
@@ -1,6 +1,6 @@
 import numpy as np
-from runepy.world import RegionManager
-from runepy.world import Region
+from runepy.world.manager import RegionManager
+from runepy.world.region import Region
 from constants import REGION_SIZE
