+++ tests/test_texture_editor.py
@@ -1,5 +1,5 @@
 import types
-from runepy.world import World
+from runepy.world.world import World
 from runepy.map_editor import MapEditor
 
 class StubFrame:
