+++ tests/test_tiledata.py
@@ -1,4 +1,4 @@
-from runepy.world import TileData
+from runepy.world.world import TileData
 
 def test_tiledata_round_trip_custom_props():
     tile = TileData(walkable=False, description="foo", properties={"cost": 5})
